+++ src/tf_object_detection_node.py
@@ -1,28 +1,20 @@
 #!/usr/bin/env python
-from __future__ import print_function # TODO is this needed
 
 import sys
 import rospy
 import object_detection_lib
-from tf_object_detection.msg import detection_results
-from std_msgs.msg import Empty
-from sensor_msgs.msg import Image
+import actionlib
+from tf_object_detection.msg import scan_for_objectsAction, scan_for_objectsGoal, scan_for_objectsResult
+from sensor_msgs.msg import CompressedImage
 from cv_bridge import CvBridge, CvBridgeError
 
 class ObjectDetectionNode:
     def __init__(self):
         self.__bridge = CvBridge()
         # Publisher to publish update image
-        self.__image_pub = rospy.Publisher("tf_object_detection_node/adjusted_image", Image, queue_size=1)
-        # Publisher to publish the result
-        self.__result_pub = rospy.Publisher("tf_object_detection_node/result", detection_results, queue_size=1)
-        # Subscribe to topic which will kick off object detection in the next image
-        self.__command_sub = rospy.Subscriber("tf_object_detection_node/start", Empty, self.StartCallback)
+        self.__image_pub = rospy.Publisher('tf_object_detection_node/image/compressed', CompressedImage, queue_size=1)
         # Subscribe to the topic which will supply the image fom the camera
-        self.__image_sub = rospy.Subscriber("camera/image/raw",Image, self.Imagecallback)
-
-        # Flag to indicate that we have been requested to use the next image
-        self.__scan_next = False
+        self.__image_sub = rospy.Subscriber('raspicam_node/image/compressed', CompressedImage, self.Imagecallback)
 
         # Read the path for models/research/object_detection directory from the parameter server or use this default
         object_detection_path = rospy.get_param('/object_detection/path', '/home/ubuntu/git/models/research/object_detection')
@@ -33,31 +25,34 @@ class ObjectDetectionNode:
         # Create the object_detection_lib class instance
         self.__odc = object_detection_lib.ObjectDetection(object_detection_path, confidence_level)
 
-    # Callback for start command message
-    def StartCallback(self, data):
-        # Indicate to use the next image for the scan
-        self.__scan_next = True
+        # Create the Action server
+        self.__as = actionlib.SimpleActionServer('object_detection', scan_for_objectsAction, self.do_action, False)
+        self.__as.start()
 
     # Callback for new image received
     def Imagecallback(self, data):
-        if self.__scan_next == True:
-            self.__scan_next = False
-            # Convert the ROS image to an OpenCV image
-            image = self.__bridge.imgmsg_to_cv2(data, "bgr8")
+        # Each time we receive an image we store it in case we are asked to scan it
+        self.__current_image = data
 
-            # The supplied image will be modified if known objects are detected
-            object_names_detected = self.__odc.scan_for_objects(image)
+    def do_action(self, goal):
+        # Scan the current image for objects
+        # Convert the ROS compressed image to an OpenCV image
+        image = self.__bridge.compressed_imgmsg_to_cv2(self.__current_image)
 
-            # publish the image, it may have been modified
-            try:
-                self.__image_pub.publish(self.__bridge.cv2_to_imgmsg(image, "bgr8"))
-            except CvBridgeError as e:
-                print(e)
+        # The supplied image will be modified if known objects are detected
+        object_names_detected = self.__odc.scan_for_objects(image)
 
-            # Publish names of objects detected
-            result = detection_results()
-            result.names_detected = object_names_detected
-            self.__result_pub.publish(result)
+        # publish the image, it may have been modified
+        try:
+            self.__image_pub.publish(self.__bridge.cv2_to_compressed_imgmsg(image))
+        except CvBridgeError as e:
+            print(e)
+
+        # Set result for the action
+        result = scan_for_objectsResult()
+        result.names_detected = object_names_detected
+        self.__as.set_succeeded(result) 
+
 
 def main(args):
     rospy.init_node('tf_object_detection_node', anonymous=False)