+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package tf_object_detection
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+Unreleased
+------------------
+* Updated to use compressed images
+* Updated to use action to start the object detection
+
 0.1.0 (2018-12-01)
 ------------------
 * First formal release of the package